move contact generator into sky130 impl block
